renaming moveVideosToFolder to move-videos-to-folder
